Combine DPC plugins into a single file
